Try factoring get_reference_data to a separate file
This could sort out import loops if it works!

commit actual private module to vc
