+++ euphonic/data/isotopes.py
@@ -492,137 +492,3 @@ class CsvData(IsotopeData):
 
 
 sears_1992 = CsvData(files(euphonic.data) / 'sears-1992.csv', {})
-
-
-def get_reference_data(
-    collection: str = 'Sears1992',
-    physical_property: str = 'coherent_scattering_length',
-) -> dict[str, Quantity]:
-    """
-    Get physical data as a dict of (possibly-complex) floats from reference
-    data.
-
-    Each "collection" refers to a JSON file which may contain any set of
-    properties, indexed by physical_property.
-
-    Properties are stored in JSON files, encoding a single dictionary with the
-    structure::
-
-      {"metadata1": "metadata1 text", "metadata2": ...,
-       "physical_properties": {"property1": {"__units__": "unit_str",
-                                             "H": H_property1_value,
-                                             "He": He_property1_value,
-                                             "Li": {"__complex__": true,
-                                                    "real": Li_property1_real,
-                                                    "imag": Li_property1_imag},
-                                             "Nh": None,
-                                             ...},
-                               "property2": ...}}
-
-    Parameters
-    ----------
-    collection
-        Identifier of data file; this may be an inbuilt data set ("Sears1992"
-        or "BlueBook") or a path to a JSON file (e.g. "./my_custom_data.json").
-
-    physical_property
-        The name of the property for which data should be extracted. This must
-        match an entry of "physical_properties" in the data file.
-
-    Returns
-    -------
-    dict[str, Quantity]
-        Requested data as a dict with string keys and (possibly-complex)
-        float Quantity values. String or None items of the original data file
-        will be omitted.
-
-    """
-
-    # Avoid import loop; the problem will go away when this function is removed
-    from euphonic.util import _deprecation_warn, comma_join, format_error
-
-    _deprecation_warn(
-        'get_reference_data', 'IsotopeData.get_array', stacklevel=3
-    )
-
-    data = _get_all_dicts_from_json(collection)
-
-    if physical_property not in data:
-        msg = format_error(
-            (
-                f'No such collection {collection!r} '
-                f'with property {physical_property!r}.'
-            ),
-            fix=(
-                f'Available properties for this collection:{comma_join(data)}.'
-            ),
-        )
-        raise ValueError(msg)
-
-    return data[physical_property]
-
-
-def _get_all_dicts_from_json(
-    collection: str = 'Sears1992',
-) -> dict[str, dict[str, Quantity]]:
-    from euphonic.util import comma_join, format_error  # Avoid import loop
-
-    _reference_data_files = {
-        'Sears1992': 'sears-1992.json',
-        'BlueBook': 'bluebook.json',
-    }
-
-    def custom_decode(dct):
-        if '__complex__' in dct:
-            return complex(dct['real'], dct['imag'])
-        return dct
-
-    if filename := _reference_data_files.get(collection):
-        file_path = files(euphonic.data) / filename
-    else:
-        file_path = Path(collection)
-
-    if not file_path.is_file():
-        msg = format_error(
-            f'No data files known for collection "{collection}".',
-            fix=f'Available collections: {comma_join(_reference_data_files)}.',
-        )
-        raise ValueError(msg)
-
-    with file_path.open() as fd:
-        file_data = json.load(fd, object_hook=custom_decode)
-
-    if 'physical_property' not in file_data:
-        msg = format_error(
-            'Data file does not contain required key "physical_property".',
-            fix='Ensure file is formatted correctly.',
-        )
-        raise AttributeError(msg)
-
-    result: dict[str[dict[str, Quantity]]] = {}
-    for physical_property, data in file_data['physical_property'].items():
-        unit_str = data.get('__units__')
-
-        no_units_msg = format_error(
-            f'No units in file ({filename}).',
-            fix='Ensure file specifies dimensions with "__units__" metadata.',
-        )
-        if unit_str is None:
-            raise ValueError(no_units_msg)
-
-        try:
-            unit = ureg(unit_str)
-        except UndefinedUnitError as exc:
-            msg = format_error(
-                f'Unsupported units ({unit_str}) from data file "{filename}".',
-                fix='Ensure units are supported by Euphonic unit register.',
-            )
-            raise ValueError(msg) from exc
-
-        result[physical_property] = {
-            key: Quantity(value, unit)
-            for key, value in data.items()
-            if isinstance(value, (float, complex))
-        }
-
-    return result
